+++ app/src/main/java/com/example/shopify/feature/MainActivity.kt
@@ -1,10 +1,13 @@
-package com.example.shopify.ui
+package com.example.shopify.feature
 
 import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import androidx.compose.runtime.Composable
+import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.navigation.compose.rememberNavController
+import com.example.shopify.feature.ShopifyGraph
+import com.example.shopify.feature.navigation_bar.productDetails.view.ProductDetailsScreen
 import com.example.shopify.theme.ShopifyTheme
 import dagger.hilt.android.AndroidEntryPoint
 
@@ -18,7 +21,7 @@ class MainActivity : ComponentActivity() {
                 // A surface container using the 'background' color from the theme
                 //App()
                 ProductDetailsScreen(
-                    id = "gid://shopify/Product/8312397005107",
+                    id = "gid://shopify/Product/8312390877491",
                     viewModel = hiltViewModel()
                 )
             }
