+++ app/src/main/java/com/example/shopify/feature/navigation_bar/productDetails/components/ProductDetailsCard.kt
@@ -1,4 +1,4 @@
-package com.example.shopify.ui.screen.productDetails.components
+package com.example.shopify.feature.navigation_bar.productDetails.components
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Arrangement
@@ -39,8 +39,8 @@ import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import coil.compose.SubcomposeAsyncImage
 import com.example.shopify.R
-import com.example.shopify.ui.theme.Gray
-import com.example.shopify.ui.theme.shopifyColors
+import com.example.shopify.theme.Gray
+import com.example.shopify.theme.shopifyColors
 import com.example.shopify.utils.shopifyLoading
 
 @Composable
@@ -52,7 +52,8 @@ fun ProductDetailsCard(
     price:String,
     realPrice:String,
     discountPercent:String,
-    quantity:String
+    quantity:String,
+    isLowStock:Boolean
 ) {
     Column(
         modifier = Modifier
@@ -160,13 +161,16 @@ fun ProductDetailsCard(
                         fontWeight = FontWeight.Medium
                     )
                 }
-                Spacer(modifier = Modifier.height(8.dp))
-                Text(
-                    text = stringResource(R.string.low_stock_only_4_left,quantity),
-                    color = Color.Red,
-                    style = MaterialTheme.typography.labelLarge,
-                    fontWeight = FontWeight.Normal,
-                )
+                if (isLowStock){
+                    Spacer(modifier = Modifier.height(8.dp))
+                    Text(
+                        text = stringResource(R.string.low_stock_only_4_left,quantity),
+                        color = Color.Red,
+                        style = MaterialTheme.typography.labelLarge,
+                        fontWeight = FontWeight.Normal,
+                    )
+                }
+
             }
 
             Column(
@@ -205,6 +209,7 @@ fun PreviewProductDetailsCard() {
         price = "172.00",
         realPrice = "249.00",
         discountPercent = "30%",
-        quantity = "4"
+        quantity = "4",
+        isLowStock = false
     )
 }
