+++ app/src/main/java/com/example/shopify/feature/navigation_bar/productDetails/components/ProductSnackBar.kt
@@ -1,4 +1,4 @@
-package com.example.shopify.ui.screen.productDetails.components
+package com.example.shopify.feature.navigation_bar.productDetails.components
 
 import androidx.compose.animation.AnimatedVisibility
 import androidx.compose.animation.expandVertically
@@ -32,9 +32,9 @@ import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.unit.dp
 import com.example.shopify.R
-import com.example.shopify.ui.screen.common.component.ShopifyOutlinedButton
-import com.example.shopify.ui.screen.common.component.ShopifyOutlinedButtonState
-import com.example.shopify.ui.theme.shopifyColors
+import com.example.shopify.feature.common.component.ShopifyOutlinedButton
+import com.example.shopify.feature.common.component.ShopifyOutlinedButtonState
+import com.example.shopify.theme.shopifyColors
 import com.example.shopify.utils.shopifyLoading
 
 @Composable
