+++ app/src/main/java/com/example/shopify/feature/navigation_bar/productDetails/components/ProductTopBar.kt
@@ -1,4 +1,4 @@
-package com.example.shopify.ui.screen.productDetails.components
+package com.example.shopify.feature.navigation_bar.productDetails.components
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Column
@@ -27,8 +27,9 @@ import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import com.example.shopify.R
-import com.example.shopify.ui.screen.common.NamedTopAppBar
-import com.example.shopify.ui.theme.shopifyColors
+import com.example.shopify.feature.common.NamedTopAppBar
+import com.example.shopify.theme.shopifyColors
+
 
 @Composable
 fun ProductTopBar(back:() -> Unit){
