+++ app/src/main/java/com/example/shopify/feature/navigation_bar/productDetails/components/VendorCard.kt
@@ -1,4 +1,4 @@
-package com.example.shopify.ui.screen.productDetails.components
+package com.example.shopify.feature.navigation_bar.productDetails.components
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.border
@@ -25,7 +25,7 @@ import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.unit.dp
 import com.example.shopify.R
-import com.example.shopify.ui.theme.shopifyColors
+import com.example.shopify.theme.shopifyColors
 
 @Composable
 fun VendorCard(
