+++ app/src/main/java/com/example/shopify/feature/navigation_bar/productDetails/view/AddToCardState.kt
@@ -1,13 +1,14 @@
-package com.example.shopify.ui.screen.productDetails.view
+package com.example.shopify.feature.navigation_bar.productDetails.view
 
-import com.example.shopify.ui.screen.productDetails.model.Price
+import com.example.shopify.feature.navigation_bar.productDetails.model.Price
 
 data class AddToCardState (
     val selectedQuantity:Int = 1,
     val isOpened:Boolean = false,
     val expandBottomSheet:Boolean = false,
     val isAdded:Boolean = false,
-    val totalCartPrice:Price = Price("",""),
+    val availableQuantity:Int = 0,
+    val totalCartPrice: Price = Price("",""),
     val sendSelectedQuantity:(Int) ->Unit,
     val openQuantity:()->Unit,
     val closeQuantity:()->Unit,
