+++ app/src/main/java/com/example/shopify/feature/navigation_bar/productDetails/view/VariantsState.kt
@@ -1,0 +1,10 @@
+package com.example.shopify.feature.navigation_bar.productDetails.view
+
+import com.example.shopify.feature.navigation_bar.productDetails.model.VariantItem
+
+data class VariantsState(
+    val variants:List<VariantItem> = listOf(),
+    val selectedVariant:Int = 1,
+    val isLowStock:Boolean = false,
+    val selectVariant:(Int)->Unit
+)
