+++ app/src/main/java/com/example/shopify/helpers/shopify/query_generator/ShopifyQueryGenerator.kt
@@ -15,4 +15,6 @@ interface ShopifyQueryGenerator {
         input: Storefront.CreditCardPaymentInputV2
     ): Storefront.MutationQuery
     fun generateUpdateCheckoutReadyQuery(paymentId: ID): Storefront.QueryRootQuery
+    fun generateProductDetailsQuery(id: String): Storefront.QueryRootQuery
+
 }
