--- app/src/main/java/com/example/shopify/model/repository/ShopifyRepository.kt
@@ -1,21 +1,0 @@
-package com.example.shopify.feature.navigation_bar.model.repository
-
-import com.example.shopify.helpers.Resource
-import com.example.shopify.feature.auth.screens.login.model.SignInUserInfo
-import com.example.shopify.feature.auth.screens.login.model.SignInUserInfoResult
-import com.example.shopify.feature.auth.screens.registration.model.SignUpUserInfo
-import com.example.shopify.feature.auth.screens.registration.model.SignUpUserResponseInfo
-import com.example.shopify.feature.navigation_bar.cart.model.Cart
-import com.example.shopify.feature.navigation_bar.home.screen.home.model.Brand
-import kotlinx.coroutines.flow.Flow
-
-interface ShopifyRepository {
-    fun signUp(userInfo: SignUpUserInfo): Flow<Resource<SignUpUserResponseInfo>>
-    fun signIn(userInfo: SignInUserInfo): Flow<Resource<SignInUserInfoResult>>
-    suspend fun saveUserInfo(userResponseInfo: SignInUserInfoResult)
-    fun getUserInfo(): Flow<SignInUserInfoResult>
-    fun isLoggedIn(): Flow<Boolean>
-    fun getBrands(): Flow<Resource<List<Brand>?>>
-    fun getCart():Flow<Resource<Cart>>
-    fun getProductDetailsByID(id: String) : Flow<Resource<Product>>
-}
