--- app/src/main/java/com/example/shopify/model/repository/ShopifyRepositoryImpl.kt
@@ -1,97 +1,0 @@
-package com.example.shopify.feature.navigation_bar.model.repository
-
-import com.example.shopify.helpers.Resource
-import com.example.shopify.helpers.shopify.query_generator.ShopifyQueryGenerator
-import com.example.shopify.feature.navigation_bar.model.local.ShopifyDataStoreManager
-import com.example.shopify.helpers.shopify.mapper.ShopifyMapper
-import com.example.shopify.feature.auth.screens.login.model.SignInUserInfo
-import com.example.shopify.feature.auth.screens.login.model.SignInUserInfoResult
-import com.example.shopify.feature.auth.screens.registration.model.SignUpUserInfo
-import com.example.shopify.feature.auth.screens.registration.model.SignUpUserResponseInfo
-import com.example.shopify.feature.navigation_bar.cart.model.Cart
-import com.example.shopify.feature.navigation_bar.home.screen.home.model.Brand
-import com.example.shopify.utils.enqueue
-import com.example.shopify.utils.mapResource
-import com.shopify.buy3.GraphCallResult
-import com.shopify.buy3.GraphClient
-import com.shopify.buy3.Storefront
-import kotlinx.coroutines.CoroutineDispatcher
-import kotlinx.coroutines.flow.Flow
-import kotlinx.coroutines.flow.flow
-import kotlinx.coroutines.flow.flowOn
-import kotlinx.coroutines.flow.map
-import javax.inject.Inject
-
-
-
-class ShopifyRepositoryImpl @Inject constructor(
-    private val graphClient: GraphClient,
-    private val queryGenerator: ShopifyQueryGenerator,
-    private val mapper: ShopifyMapper,
-    private val dataStoreManager: ShopifyDataStoreManager,
-    private val defaultDispatcher: CoroutineDispatcher
-) : ShopifyRepository {
-
-    override fun signUp(userInfo: SignUpUserInfo): Flow<Resource<SignUpUserResponseInfo>> {
-        return queryGenerator.generateSingUpQuery(userInfo)
-            .enqueue()
-            .mapResource(mapper::map)
-    }
-
-    override fun signIn(userInfo: SignInUserInfo): Flow<Resource<SignInUserInfoResult>> {
-        return queryGenerator.generateSingInQuery(userInfo)
-            .enqueue()
-            .mapResource { response -> mapper.mapToSignInResponse(response, userInfo) }
-    }
-
-    override suspend fun saveUserInfo(userResponseInfo: SignInUserInfoResult) =
-        dataStoreManager.saveUserInfo(userResponseInfo)
-
-
-    override fun getUserInfo(): Flow<SignInUserInfoResult> =
-        dataStoreManager.getUserInfo()
-
-    override fun getProductDetailsByID(id:String) : Flow<Resource<Product>> {
-        val query = queryGenerator.generateProductDetailsQuery(id)
-        return query.enqueue().mapResource(mapper::mapToProduct)
-    }
-
-
-    override fun isLoggedIn(): Flow<Boolean> =
-        dataStoreManager.getAccessToken()
-            .map { it != null }
-            .flowOn(defaultDispatcher)
-
-    override fun getBrands(): Flow<Resource<List<Brand>?>> {
-        val query = queryGenerator.generateBrandQuery()
-        return query!!.enqueue().mapResource(mapper::mapToBrandResponse)
-    }
-
-    override fun getCart(): Flow<Resource<Cart>> = flow {
-    }
-
-
-    private fun Storefront.QueryRootQuery.enqueue() =
-        graphClient.enqueue(this).map { result ->
-            when (result) {
-                is GraphCallResult.Success ->
-                    Resource.Success(result.response)
-
-                is GraphCallResult.Failure ->
-                    Resource.Error(mapper.map(result.error))
-            }
-        }.applyDispatcher()
-
-    private fun Storefront.MutationQuery.enqueue() =
-        graphClient.enqueue(this).map { result ->
-            when (result) {
-                is GraphCallResult.Success ->
-                    Resource.Success(result.response)
-
-                is GraphCallResult.Failure ->
-                    Resource.Error(mapper.map(result.error))
-            }
-        }.applyDispatcher()
-
-    private fun <T> Flow<T>.applyDispatcher() = this.flowOn(defaultDispatcher)
-}
